Add draw_arrangement file
`draw_arrangement.py` contains a class to draw an arrangment of lines
gives a list of lines, storing the result of the arrangement in a BPS.
this allows us to move this part of the animation out of the draw zone
code so that the draw zone code only finds the zone and draws it.
